+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-## 0.1.0 : 11.08.2025
+## 0.1.0 : 12.08.2025
 
+- **Feature**: Updated README and refactored functions.
 - **Feature**: Added test file and save output to CSV as a single file.
 - **Feature**: Added transformations and hash utils.
 - **Feature**: Added main script, config and data files.